+++ probe/src/cgo/kindling.cpp
@@ -31,6 +31,8 @@ int64_t debug_pid = 0;
 int64_t debug_tid = 0;
 char* traceId = new char[128];
 char* isEnter = new char[16];
+char* protocol = new char[32];
+char* url = new char[256];
 char* start_time_char = new char[32];
 char* end_time_char = new char[32];
 char* tid_char = new char[32];
@@ -465,6 +467,8 @@ void parse_xtid(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
   int val_offset = 0;
   int tmp_offset = 0;
   int traceId_offset = 0;
+  int protocol_offset = 0;
+  int url_offset = 0;
   for (int i = 8; i < data_param.m_len; i++) {
     if (data_val[i] == '!') {
       if (val_offset == 0) {
@@ -472,6 +476,12 @@ void parse_xtid(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
         traceId_offset = tmp_offset;
       } else if (val_offset == 1) {
         isEnter[tmp_offset] = '\0';
+      } else if (val_offset == 2) {
+        protocol[tmp_offset] = '\0';
+        protocol_offset = tmp_offset;
+      } else if (val_offset == 3) {
+        url[tmp_offset] = '\0';
+        url_offset = tmp_offset;
         break;
       }
       tmp_offset = 0;
@@ -482,7 +492,12 @@ void parse_xtid(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
       traceId[tmp_offset] = data_val[i];
     } else if (val_offset == 1) {
       isEnter[tmp_offset] = data_val[i];
+    } else if (val_offset == 2) {
+      protocol[tmp_offset] = data_val[i];
+    } else if (val_offset == 3) {
+      url[tmp_offset] = data_val[i];
     }
+
     tmp_offset++;
   }
   p_kindling_event->timestamp = s_evt->get_ts();
@@ -496,6 +511,16 @@ void parse_xtid(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
   p_kindling_event->userAttributes[userAttNumber].valueType = CHARBUF;
   p_kindling_event->userAttributes[userAttNumber].len = 1;
   userAttNumber++;
+  strcpy(p_kindling_event->userAttributes[userAttNumber].key, "protocol");
+  memcpy(p_kindling_event->userAttributes[userAttNumber].value, protocol, protocol_offset);
+  p_kindling_event->userAttributes[userAttNumber].valueType = CHARBUF;
+  p_kindling_event->userAttributes[userAttNumber].len = protocol_offset;
+  userAttNumber++;
+  strcpy(p_kindling_event->userAttributes[userAttNumber].key, "url");
+  memcpy(p_kindling_event->userAttributes[userAttNumber].value, url, url_offset);
+  p_kindling_event->userAttributes[userAttNumber].valueType = CHARBUF;
+  p_kindling_event->userAttributes[userAttNumber].len = url_offset;
+  userAttNumber++;
   strcpy(p_kindling_event->name, "apm_trace_id_event");
   p_kindling_event->context.tinfo.tid = threadInfo->m_tid;
   map<uint64_t, char*>::iterator key =
@@ -506,17 +531,28 @@ void parse_xtid(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
     strcpy(p_kindling_event->context.tinfo.comm, (char*)threadInfo->m_comm.data());
   }
   p_kindling_event->context.tinfo.pid = threadInfo->m_pid;
+  strcpy(p_kindling_event->context.tinfo.containerId, (char*)threadInfo->m_container_id.data());
   p_kindling_event->paramsNumber = userAttNumber;
 }
 
 void parse_span(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
                 kindling_event_t_for_go* p_kindling_event, sinsp_threadinfo* threadInfo,
                 uint16_t& userAttNumber) {
+  if (data_param.m_len < 10) {
+    return;
+  }
   int val_offset = 0;
   int tmp_offset = 0;
   int span_offset = 0;
   int traceId_offset = 0;
-  for (int i = 8; i < data_param.m_len; i++) {
+
+  bool version_new = false;
+  int fromIndex = 8;
+  if (data_val[8] == '1' && data_val[9] == '!') {
+    version_new = true;
+    fromIndex = 10;
+  }
+  for (int i = fromIndex; i < data_param.m_len; i++) {
     if (data_val[i] == '!') {
       if (val_offset == 0) {
         duration_char[tmp_offset] = '\0';
@@ -541,7 +577,13 @@ void parse_span(sinsp_evt* s_evt, char* data_val, sinsp_evt_param data_param,
     }
     tmp_offset++;
   }
-  p_kindling_event->timestamp = s_evt->get_ts() - atol(duration_char);
+  if (version_new) {
+    // StartTime
+    p_kindling_event->timestamp = atol(duration_char);
+  } else {
+    // EndTime - Duration
+    p_kindling_event->timestamp = s_evt->get_ts() - atol(duration_char);
+  }
   strcpy(p_kindling_event->userAttributes[userAttNumber].key, "end_time");
   memcpy(p_kindling_event->userAttributes[userAttNumber].value, to_string(s_evt->get_ts()).data(),
          19);
@@ -916,6 +958,83 @@ void attach_pid(char* pid, bool is_new_start, bool is_attach, bool is_all_attach
   if (!is_ps && !is_attach) {
     cout << "------end detach for pid " << pid << "------" << endl;
   }
+}
+
+void attach_agent(int64_t pid, char* error_message, bool is_attach) {
+  char result_buf[1024], command[1024];
+  string attach_command_prefix;
+  if (is_attach) {
+    attach_command_prefix = "./async-profiler/jattach.sh start ";
+  } else {
+    attach_command_prefix = "./async-profiler/jattach.sh stop ";
+  }
+  attach_command_prefix.append(std::to_string(pid));
+  strcpy(command, attach_command_prefix.c_str());
+
+  FILE* fp;
+  fp = popen(command, "r");
+  if (NULL == fp) {
+    perror("popen execute failed!\n");
+    strcpy(error_message, "popen execute failed");
+    return;
+  }
+  if (is_attach) {
+    cout << "------"
+         << " start attach agent for pid " << pid << "------" << endl;
+  } else {
+    cout << "------"
+         << " start detach agent for pid " << pid << "------" << endl;
+  }
+
+  char* error_msg;
+  while (fgets(result_buf, sizeof(result_buf), fp) != NULL) {
+    if ('\n' == result_buf[strlen(result_buf) - 1]) {
+      result_buf[strlen(result_buf) - 1] = '\0';
+    }
+    error_msg = strstr(result_buf, "[ERROR] ");
+    if (error_msg) {
+      strcpy(error_message, error_msg);
+    }
+    printf("%s\r\n", result_buf);
+  }
+
+  int rc = pclose(fp);
+  if (-1 == rc) {
+    perror("close command fp failed!\n");
+    strcpy(error_message, "close command fp failed");
+    return;
+  } else {
+    printf("command:【%s】command process status:【%d】command return value:【%d】\r\n", command,
+           rc, WEXITSTATUS(rc));
+  }
+
+  if (is_attach) {
+    cout << "------end attach agent for pid " << pid << "------" << endl;
+  } else {
+    cout << "------end detach agent for pid " << pid << "------" << endl;
+  }
+}
+
+char* start_attach_agent(int64_t pid) {
+  char* error_message = (char*) malloc(1024 * sizeof(char));
+  error_message[0] = '\0';
+  if (!inspector) {
+    strcpy(error_message, "Please start profile first");
+  } else {
+    attach_agent(pid, error_message, true);
+  }
+  return error_message;
+}
+
+char* stop_attach_agent(int64_t pid) {
+  char* error_message = (char*) malloc(1024 * sizeof(char));
+  error_message[0] = '\0';
+  if (!inspector) {
+    strcpy(error_message, "Please start profile first");
+  } else {
+    attach_agent(pid, error_message, false);
+  }
+  return error_message;
 }
 
 int start_profile() {